Renamed BaseApp to App
